pyproject.toml
[project]
name = "agent-marketplace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.4",
    "httpx>=0.28.1",
    "openai>=2.17.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.1",
    "sqlalchemy==2.0.*",
    "uvicorn>=0.40.0",
]
